legal_agreement_analyzer/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart

# LangChain and AI
langchain
langchain-google-genai
langchain-community

# PDF Processing
PyPDF2

# Vector Store & Embeddings
chromadb
sentence-transformers
